feat: 주문 결제에 대한 이벤트 리스너 적용 및 DTO 구조 개선
## 이벤트 기반으로 주문 트랜잭션이 커밋된 후 비동기로 PG 결제요청 흐름 구현

1. 이벤트 발행 PgPaymentStrategy.java
```java
@Override
public OrderModel processPayment(PaymentCommand.ProcessPaymentContext context) {
    // sth...

    eventPublisher.publishEvent(new PgPaymentRequestedEvent(paymentCommand));

    return pendingOrder;
}
```

2. 이벤트 수신 시점

```java
// new PgPaymentRequestedEvent(paymentCommand) 발행 후
// 그에 대한 트랜잭션이 정상적으로 커밋되며 끝날 때
@TransactionalEventListener(phase = AFTER_COMMIT) // AFTER_COMMIT 리스너
```

3. 이벤트 수신 후 처리
```java
/**
 * 주문 트랜잭션이 커밋된 후 비동기로 PG 결제 요청
 */
@Async("pgExecutor")
@TransactionalEventListener(phase = AFTER_COMMIT)
public void handlePgPaymentRequested(PgPaymentRequestedEvent event) {

    log.info("[PG][EVENT] PgPaymentRequestedEvent received. userId={}, orderId={}",
            event.command().userId(), event.command().orderId());

    // 실제 PG 호출은 서비스에 위임
    pgPaymentService.requestPaymentForOrder(event.command());
}
```
