+++ apps/commerce-api/src/main/java/com/loopers/application/payment/event/PgPaymentRequestedEvent.java
@@ -1,0 +1,8 @@
+package com.loopers.application.payment.event;
+
+import com.loopers.application.payment.dto.PgPaymentCommand;
+
+public record PgPaymentRequestedEvent(
+        PgPaymentCommand command
+) {
+}
